requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
aiomysql
httpx
authlib
python-jose[cryptography]
requests
pytest
pytest-asyncio
